+++ src/main/gen/cruelteam/plugins/mnlang/psi/MnlangTypes.java
@@ -11,11 +11,13 @@ public interface MnlangTypes {
   IElementType ENTRY = new MnlangElementType("ENTRY");
 
   IElementType COMMENT = new MnlangTokenType("COMMENT");
+  IElementType DOC_COMMENT = new MnlangTokenType("DOC_COMMENT");
   IElementType ESCAPE = new MnlangTokenType("ESCAPE");
   IElementType KEY = new MnlangTokenType("KEY");
   IElementType PREFIX = new MnlangTokenType("PREFIX");
   IElementType SEPARATOR = new MnlangTokenType("SEPARATOR");
   IElementType VALUE = new MnlangTokenType("VALUE");
+  IElementType WHITE_SPACE = new MnlangTokenType("WHITE_SPACE");
 
   class Factory {
     public static PsiElement createElement(ASTNode node) {